refactor: Rename capitalize module to strings
