+++ src/lib/strings.ts
